Fix inline tool protocol false positives
